requirements.txt
openai-whisper
torch
pydub